refactor(client): Move LinkButton component to its own file for better organization
